refactor(engine): modularize services and fix dice fairness
